engine: extract esm/planets.js (but defer the 3D planet-drawing code for now)
